🏷️ fix(payment): 가격 필드를 int 타입으로 변경

+++ src/main/java/com/junwoo/ott/domain/payment/dto/response/OrderItemResponseDto.java
@@ -11,8 +11,8 @@ public class OrderItemResponseDto {
 
   private Long orderItemId;
   private String itemCode;
-  private Long totalPrice;
-  private Long discountingPrice;
-  private Long discountedPrice;
+  private Integer totalPrice;
+  private Integer discountingPrice;
+  private Integer discountedPrice;
 
 }
